Create 1.1.9. Fractional Knapsack Problem
The below program has a method void knapsack(). Which takes four parameters number of objects, the weight of each object, the profit corresponding to each one and the capacity of the knapsack. Write a program using a fractional knapsack algorithm to get the maximum profit.



Print the output as follows:

Sample Input and Output:
Enter the no. of objects: 6
Enter the weights and profits of each object:
1 2
4 5
8 9
4 6
5 2
3 5
Enter the capacity of knapsack:10
Maximum profit is:- 15.500000